disable-source: move command to own source file
